pyproject.toml
[tool.poetry]
name = "ska-oso-oet"
version = "6.6.0"
description = "SKA Observation Execution Tool"
authors = ["Team Buttons"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/ska-telescope/oso/ska-oso-oet"
documentation = "https://developer.skao.int/projects/ska-oso-oet/en/latest/index.html"
keywords = ["SKA OSO"]
packages = [
    { include = "ska_oso_oet", from = "src" }
]
include = [
    "src/ska_oso_oet/ska_oso_oet.ini",
    { path = 'tests' }
]
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: BSD License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
]

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'
priority = 'supplemental'

[[tool.poetry.source]]
name = 'ska-scripting-gitlab'
url = 'https://gitlab.com/api/v4/projects/22057734/packages/pypi/simple'
priority = "explicit"

[[tool.poetry.source]]
name = 'PyPI'
priority = 'primary'

# Uncomment this and the dependency below to use an unreleased ODA version
#[[tool.poetry.source]]
#name = "gitlab-ska-db-oda"
#url = "https://gitlab.com/api/v4/projects/27507143/packages/pypi/simple"

#[[tool.poetry.source]]
#name = "gitlab.ska-oso-oet-client"
#url = "https://gitlab.com/api/v4/projects/39372522/packages/pypi/simple"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "~2.2"
jsonpickle = "^2.2.0"
PyPubSub = "^4.0.3"
requests = "^2.28.1"
ska-ser-logging = "^0.4.1"
ska-ser-skuid = "^3.3.3"
tblib = "^1.7.0"
GitPython = "^3.1.27"
waitress = "^2.1.2"
# The OET depends on the PDM as we need access to SBD activities, etc., but the
# data model is evolving so fast moving that we don't want to pin to a specific
# version. The best compromise for compatibility with ska-oso-scripting, which
# also depends on the OET and ODA, is implicitly depend on whatever version of
# the PDM is pulled in by the ODA.
#ska-db-oda = { version="5.3.3+dev.cea0b078b", source="gitlab-ska-db-oda" }
connexion = {extras = ["swagger-ui"], version = "^2.14.2"}
# Having to pin some more versions that cause an issue with the default environment. Should be able
# to remove once the OET <-> scripting circular dependency is broken
prance = {version = "^23.6.21.0", extras = ["osv"]}
jsonschema = "4.17.3"
# NOTE: When updating ODA version, also update the ODA chart version in
# charts/ska-oso-oet-umbrella/Chart.yaml
ska-db-oda = "^7.3.0"
ska-oso-scripting = "^11.0.0"
fastapi = "^0.111.0"
pydantic = "^2.10.6"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
coverage = "^6.5.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mock = "^4.0.3"
pylint = "^2.15.3"
pylint-junit = "^0.3.2"
pytest = "^7.1.3"
pytest-bdd = "^6.0.1"
pytest-cov = "^4.0.0"
pytest-env = "^0.6.2"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.9.0"
requests-mock = "^1.10.0"
ska-oso-oet-client = "^1.2.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinxcontrib-openapi = "^0.8.4"
ska-ser-sphinx-theme = "^0.1.3"
recommonmark = "^0.7.1"
# required to resolve the following import issue:
# Could not import extension sphinxcontrib.openapi (exception: cannot import name 'error_string' from 'docutils.io'
mistune = "^2"

[tool.poetry.scripts]
oet = "ska_oso_oet_client.main:main"

[tool.pytest.ini_options]
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s.%(msecs)03dZ|%(levelname)s|%(threadName)s|%(funcName)s|%(filename)s#%(lineno)d||%(message)s"
log_cli_date_format = "%Y-%m-%dT%H:%M:%S"
log_file = "pytest-logs.txt"
log_file_format = "%(asctime)s.%(msecs)03dZ|%(levelname)s|%(threadName)s|%(funcName)s|%(filename)s#%(lineno)d||%(message)s"
log_file_level = "DEBUG"
env = [
    "D:ODA_URL = http://ska-db-oda-rest-test:5000/ska-oso-oet/oda/api/v3"
]

[tool.coverage.run]
parallel = true
branch = true
data_file = ".coverage/.coverage"

[tool.black]
line-length = 88
preview = true
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
